include plone5 specific yafowil font
